requirements.txt
aiohttp==3.12.15
py-cord==2.6.1
mysql-connector-python==9.4.0
python-dotenv==1.1.1
